test where 8 players sign up, go to games page, player1 goes to ccreate game page, and then they all delete their accounts

+++ app/(main)/account/page.jsx
@@ -123,6 +123,7 @@ const AccountPAge= () => {
     }
 
     const deleteAccount = async () => {
+        if(!confirm('Are you sure you want to delete your account?'))return
         try{
             const deletedUser = await deleteUserAPI(session.user.id)
             await signOut()
